Rough CSS Style completed
The basic foundation of CSS is laid out (though I got really tired at the end and ended up doing lazy naming for some of the div classes). I had to use a p class subtitle for a header since it kept conflicting with the title not aligning correctly to the left.